move tap, curry into function category
