+++ function/tap.js
