+++ logic/__tests__/unless.test.js
@@ -1,6 +1,6 @@
 const unless = require('../unless')
 
-describe('core/unless', () => {
+describe('logic/unless', () => {
   const isOdd = num => num % 2 !== 0
   const isEven = num => num % 2 === 0
   const double = num => num * 2
